Don't bridge low level controls for fixed wings
Since we have an attitude controller, it does not make sense to bridge low level controls to teams for fixed wings as the commands will be over written by the attitude controller. Instead they should publish messages to attitude controllers.

An alternative to this PR would be to have an enable/disable flag for the ttitude controller and bridge that over.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>