Fixes CI tests (#90)
* Fixes CI tests

Two tests were causing an :x: in the CI. The first was `test_spawn_usv_maxspeed`. This PR fixes `test_spawn_usv_maxspeed` by creating a seperate test world for it without the gamelogic plugin. This is because the gamelogic plugin was terminating the test before it could be completed. I also increased the test timeouts.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>

* debug sdf

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>