Adds a simple Maritime radar.
The following PR adds a simple maritime RADAR. It is by no means complete but since I dont have time on this project anymore I'll leave this here.  Bascially, the way this works is very simple. We extend the naive spiinning radar example with a simple plugin. This plugin projects the radar onto a polar coordinate frame. It combines the current lidar position and the lidar return and projects it into a polar plot like maritime radars do. The end user will receive a `ignition.msgs.Float_V`. The first bin contains the angle the radar is currently facing, the second bin will contain the resolution of the bin and the third item is the resolution of each bin. The rest of the FloatV will contain the sampled bins for each location.

## Some things to work on

* [ ] Performance. Currently the performance of the system is at 25%RTF when we load the radar simulation. There are some low hanging fruits which I have marked with TODOs that should help boost the performance. In particular we are doing a lot of copying when filling the message. Pre-allocating the message could save some compute time.
* [ ] ROS Bridge.  Currently the function in the rosbridge looks correct but for whatever reason Im unable to link against the local protobuf runtime. (The Naive RADAR bridge is doing the same for me)
* Unit/integration tests

## Stretch Goals
* Integrate rain and sea state noise in (see todos in `MaritimeRadar.cc`)

For completeness the NaiveRadar should really just populate the RadarTracks message instead of the RadarScan message.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>